DELFT3DFM-1391: When closing RGFGRID without saving, instead of a cryptic and unclear message box, we now show an information window that tells the user to save his/her work before exiting RGFGRID.
Note: This is a temporary measure! When RGFGRID has the functionality that prompts the user for saving their work when exiting and not saved, then this message box should be removed from the code.
Former-commit-id: 511a1bd0b9845b343269caf11cfa09cb2eb6418c
Former-commit-id: cfa89115549310434df9e39af9e1965fb8609d9f

+++ src/DeltaShell.Plugins.FMSuite.FlowFM.Gui/FlowFMGuiPlugin.cs
@@ -231,10 +231,11 @@ namespace DeltaShell.Plugins.FMSuite.FlowFM.Gui
                         {
                             SharpMapGisGuiPlugin.Instance.Gui.MainWindow.SetWaitCursorOn();
                         }
+                        // D3DFMIQ-16: This if-statement becomes redundant when RGFGRID asks the user to save the grid on exit (when grid was not saved)
                         if (File.Exists(model.NetFilePath) && new FileInfo(model.NetFilePath).Length == 0)
                         {
                             throw new FileFormatException(new Uri(model.NetFilePath),
-                                "invalid empty file detected. Please save your project after editing in RGFGrid.");
+                                Properties.Resources.FlowFMGuiPlugin_GetViewInfoObjects_Empty_file_detected__Changes_in_the_grid_were_not_saved);
                         }
                         if (!File.Exists(model.NetFilePath))
                         {
@@ -246,12 +247,20 @@ namespace DeltaShell.Plugins.FMSuite.FlowFM.Gui
                         {
                             var netfile = new ImportedFMNetFile(model.NetFilePath);
                             var coordinates = netfile.Grid.Vertices;
-                            if (!CoordinateSystemValidator.CanAssignCoordinateSystem(coordinates, model.CoordinateSystem))
+                            if (!CoordinateSystemValidator.CanAssignCoordinateSystem(coordinates,
+                                model.CoordinateSystem))
                             {
-                                throw new Exception("Grid coordinates are incompatible with current model coordinate system");
+                                throw new Exception(
+                                    "Grid coordinates are incompatible with current model coordinate system");
                             }
                         }
                         model.ReloadGrid(false);
+                    }
+                    // D3DFMIQ-16: This catch block becomes redundant when RGFGRID asks the user to save the grid on exit (when grid was not saved)
+                    catch (FileFormatException exception) 
+                    {
+                        MessageBox.Show(exception.Message, "Grid was not saved in RGFGRID", MessageBoxButtons.OK);
+                        model.Grid = NetFileImporter.ImportGrid(model.NetFilePath) ?? new UnstructuredGrid();
                     }
                     catch (Exception exception)
                     {
@@ -266,10 +275,7 @@ namespace DeltaShell.Plugins.FMSuite.FlowFM.Gui
                         }
                         else
                         {
-                            if (File.Exists(model.NetFilePath))
-                            {
-                                File.Delete(model.NetFilePath);
-                            }
+                            if (File.Exists(model.NetFilePath)) File.Delete(model.NetFilePath);
                             model.WriteNetFile(model.NetFilePath);
                         }
                     }
